フロントエンドの保守性と開発効率を向上させるため、大規模なリファクタリングを実施しました。
主な変更点は以下の通りです。

1.  **コンポーネントディレクトリの再編成:**
    -   `src/components/providers` ディレクトリを新設し、`theme-provider.tsx` を移動しました。
    -   `mode-toggle.tsx` を `src/components/elements` に移動しました。
    -   ファイル移動に伴い、関連するすべての `import` パスを修正しました。

2.  **ページコンポーネントの分割:**
    -   `about`, `gallery`, `series`, `destination` の各ページで、`Client.tsx` や関連コンポーネントに直接記述されていたUIロジックを、より小さく再利用可能なコンポーネントに分割しました。
    -   分割したコンポーネントは `src/components/featured/[page-name]` ディレクトリに格納し、元のファイルはそれらのコンポーネントを呼び出すコンテナとしての役割に専念するように修正しました。

このリファクタリングにより、各コンポーネントの責務が明確になり、コードベース全体の見通しが大幅に改善されました。
